🚚 rename: axios, useQuery 요청 훅 이름 변경(user가 아닌 다른 데이터도 같이 있어서 이름을 변경)
